requirements.txt
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.47.0
python-dotenv==1.0.1
